+++ agent/agent-profiler/agent-diagnostics/src/main/java/com/microsoft/applicationinsights/diagnostics/collection/libos/os/linux/cgroups/CGroupCpuUsageReader.java
@@ -4,9 +4,9 @@
 package com.microsoft.applicationinsights.diagnostics.collection.libos.os.linux.cgroups;
 
 @SuppressWarnings("checkstyle:AbbreviationAsWordInName")
-public class CGroupCPUSystemReader extends CGroupValueReader {
-  // total system CPU time (in nanoseconds) consumed by all tasks in this cgroup
-  public CGroupCPUSystemReader() {
-    super("/sys/fs/cgroup/cpu,cpuacct/cpuacct.usage_sys");
+public class CGroupCpuUsageReader extends CGroupValueReader {
+  // total CPU usage (in nanoseconds) consumed by all tasks in this cgroup
+  public CGroupCpuUsageReader() {
+    super("/sys/fs/cgroup/cpu,cpuacct/cpuacct.usage");
   }
 }