+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/httpclient/LazyHttpClient.java
@@ -97,6 +97,7 @@ public class LazyHttpClient implements HttpClient {
 
     NettyAsyncHttpClientBuilder builder = new NettyAsyncHttpClientBuilder();
     if (proxyHost != null && proxyPortNumber != null) {
+      @SuppressWarnings("AddressSelection")
       ProxyOptions proxyOptions =
           new ProxyOptions(
               ProxyOptions.Type.HTTP, new InetSocketAddress(proxyHost, proxyPortNumber));