+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/init/AiConfigCustomizer.java
@@ -139,7 +139,7 @@ public class AiConfigCustomizer implements Function<ConfigProperties, Map<String
     properties.put("otel.instrumentation.apache-httpasyncclient.enabled", "true");
     properties.put("otel.instrumentation.apache-httpclient.enabled", "true");
     properties.put("otel.instrumentation.async-http-client.enabled", "true");
-    properties.put("otel.instrumentation.executor.enabled", "true");
+    properties.put("otel.instrumentation.executors.enabled", "true");
     properties.put("otel.instrumentation.google-http-client.enabled", "true");
     properties.put("otel.instrumentation.grpc.enabled", "true");
     properties.put("otel.instrumentation.guava.enabled", "true");
@@ -282,9 +282,15 @@ public class AiConfigCustomizer implements Function<ConfigProperties, Map<String
     if (config.preview.instrumentation.vertx.enabled) {
       properties.put("otel.instrumentation.vertx.enabled", "true");
     }
+    if (config.preview.instrumentation.ktor.enabled) {
+      properties.put("otel.instrumentation.ktor.enabled", "true");
+    }
     if (config.preview.instrumentation.jaxrsAnnotations.enabled) {
       properties.put("otel.instrumentation.jaxrs-1.0.enabled", "true");
       properties.put("otel.instrumentation.jaxrs-annotations.enabled", "true");
+    }
+    if (config.preview.instrumentation.r2dbc.enabled) {
+      properties.put("otel.instrumentation.r2dbc.enabled", "true");
     }
   }