+++ agent/agent-tooling/src/test/java/com/microsoft/applicationinsights/agent/internal/perfcounter/JmxDataFetcherTest.java
@@ -18,6 +18,7 @@ import org.junit.jupiter.api.Test;
 class JmxDataFetcherTest {
 
   @SuppressWarnings({"unused", "checkstyle:AbbreviationAsWordInName"})
+  // This class name must end in MXBean (case sensitive)
   public interface StubMXBean {
     int getIntSample();