+++ smoke-tests/apps/SystemExit/src/smokeTest/java/com/microsoft/applicationinsights/smoketest/SystemExitTest.java
@@ -7,9 +7,9 @@ import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.JAVA_
 import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.JAVA_11_OPENJ9;
 import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.JAVA_17;
 import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.JAVA_17_OPENJ9;
-import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.JAVA_18_OPENJ9;
-import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.JAVA_19;
 import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.JAVA_20;
+import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.JAVA_20_OPENJ9;
+import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.JAVA_21;
 import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.JAVA_8;
 import static com.microsoft.applicationinsights.smoketest.EnvironmentValue.JAVA_8_OPENJ9;
 
@@ -60,12 +60,12 @@ abstract class SystemExitTest {
   @Environment(JAVA_17_OPENJ9)
   static class Java17OpenJ9Test extends SystemExitTest {}
 
-  @Environment(JAVA_19)
-  static class Java18Test extends SystemExitTest {}
-
-  @Environment(JAVA_18_OPENJ9)
-  static class Java18OpenJ9Test extends SystemExitTest {}
-
   @Environment(JAVA_20)
-  static class Java19Test extends SystemExitTest {}
+  static class JavaLatestTest extends SystemExitTest {}
+
+  @Environment(JAVA_20_OPENJ9)
+  static class JavaLatestOpenJ9Test extends SystemExitTest {}
+
+  @Environment(JAVA_21)
+  static class JavaPrereleaseTest extends SystemExitTest {}
 }